MT#55283 eliminate last_rtp_index
This field is duplicated with `ext_seq`, which is already kept in shared
memory and therefore preferable.

Also update all sequence number fields to 32-bit (down from 64-bit) as
there's no need for all these extra bits.

Add output seq information to /list.

+++ include/crypto.h
@@ -19,9 +19,9 @@ struct crypto_context;
 struct rtp_header;
 struct rtcp_packet;
 
-typedef int (*crypto_func_rtp)(struct crypto_context *, struct rtp_header *, str *, uint64_t);
+typedef int (*crypto_func_rtp)(struct crypto_context *, struct rtp_header *, str *, uint32_t);
 typedef int (*crypto_func_rtcp)(struct crypto_context *, struct rtcp_packet *, str *, uint64_t);
-typedef int (*hash_func_rtp)(struct crypto_context *, char *out, str *in, uint64_t);
+typedef int (*hash_func_rtp)(struct crypto_context *, char *out, str *in, uint32_t);
 typedef int (*hash_func_rtcp)(struct crypto_context *, char *out, str *in);
 typedef int (*session_key_init_func)(struct crypto_context *);
 typedef int (*session_key_cleanup_func)(struct crypto_context *);
@@ -123,12 +123,12 @@ INLINE int crypto_params_sdes_cmp(const struct crypto_params_sdes *cs, gconstpoi
 
 
 INLINE int crypto_encrypt_rtp(struct crypto_context *c, struct rtp_header *rtp,
-		str *payload, uint64_t index)
+		str *payload, uint32_t index)
 {
 	return c->params.crypto_suite->encrypt_rtp(c, rtp, payload, index);
 }
 INLINE int crypto_decrypt_rtp(struct crypto_context *c, struct rtp_header *rtp,
-		str *payload, uint64_t index)
+		str *payload, uint32_t index)
 {
 	return c->params.crypto_suite->decrypt_rtp(c, rtp, payload, index);
 }
